fix: moved hook to query
